+++ spec/adhearsion/call_controller/output_spec.rb
@@ -8,10 +8,30 @@ module Adhearsion
       include CallControllerTestHelpers
 
       describe "#new_play" do
-        it "should return a Play component targetted at the current controller" do
-          play = controller.new_play
-          play.should be_a Output::Play
-          play.controller.should be controller
+        context "as a standard object" do
+          it "should return a Play component targetted at the current controller" do
+            play = controller.new_play
+            play.should be_a Output::Play
+            play.controller.should be controller
+          end
+
+          it "should not be an actor" do
+            play = controller.new_play
+            play.should_not be_a Celluloid
+          end
+        end
+
+        context "as an actor" do
+          it "should return a Play component targetted at the current controller" do
+            play = controller.new_play true
+            play.should be_a Output::Play
+            play.controller.should be controller
+          end
+
+          it "should be an actor" do
+            play = controller.new_play true
+            play.should be_a Celluloid
+          end
         end
       end
 
@@ -29,6 +49,26 @@ module Adhearsion
       describe "#speak" do
         it "should be an alias for #say" do
           subject.method(:speak).should be == subject.method(:say)
+        end
+      end
+
+      describe "asynchronous" do
+        [:say!, :play!, :play_audio!, :play_time!, :play_numeric!, :interruptible_play!].each do |method_name|
+          describe "##{method_name}" do
+            it "should delegate to a new play actor and return it" do
+              play_component = flexmock Output::Play.new(controller)
+              controller.should_receive(:new_play).with(true).and_return play_component
+              play_component.should_receive(method_name).once.with 'foo'
+              val = controller.send method_name, 'foo'
+              val.should be play_component
+            end
+          end
+        end
+
+        describe "#speak!" do
+          it "should be an alias for #say!" do
+            subject.method(:speak!).should be == subject.method(:say!)
+          end
         end
       end
     end